+++ navocr/detector_vino.py
