+++ setup.py
@@ -18,7 +18,7 @@ setup(
     license='Apache-2.0',
     entry_points={
         'console_scripts': [
-            'navocr_with_ocr_node = navocr.navocr_with_ocr_node:main',
+            'navocr_with_ocr_node = navocr.navocr_with_ocr_node_vino:main',
         ],
     },
 )
